Style: styled-components 파일 분리
