+++ src/pages/PageRouter.jsx
@@ -4,10 +4,10 @@ import EmailJoin from '../components/join/EmailJoin';
 import Login from '../components/login/Login';
 import FollowersList from './FollowersList';
 import FollowingsList from './FollowingsList';
-import Home from './Home';
+import Home from './home/Home';
 import NonFollowing from './NonFollowing';
-import NotFound from './NotFound';
-import Search from './Search';
+import NotFound from './notFound/NotFound';
+import Search from './search/Search';
 import PostDetail from './postDetail/PostDetail';
 import UserProfile from './userprofile/UserProfile';
 import PostUpload from './postUpload/PostUpload';
