+++ src/pages/search/Styled.js
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const View = styled.div`
+    width: 450px;
+    height: 100%;
+    position: absolute;
+    overflow: hidden;
+`;
+
+export const ScrollBlind = styled.div`
+    width: 480px;
+    height: 100%;
+    overflow-y: scroll;
+    background-color: white;
+`;
+
+export const Ul = styled.ul`
+    padding: 60px 0;
+    margin: 0 16px;
+    height: 100vh;
+`;
